rename Cardano.Wallet.Shelley.Pools to Cardano.Wallet.Pools
